Add files via upload
removed cooks test, and the chiquare test i dont think that is needed? summarized the model not the data. box plot just the age, alco, and tabaco usage. removed the model where it was just the cancer cases count, removed the 2-way anova i dont think its needed. rearrange the tukey test to just have all the age stuff together and the Con. interval plot before our inference. same with alch. and added an interpretation for the 3-way ANOVA.